agent-1: extract changelog truth boundary
